Fix notification issue and reset timer to 30 seconds
