+++ backend/matching-service/consumer/worker/worker.go
@@ -39,7 +39,7 @@ func SpinMQConsumer(criteria utils.MatchCriteria) {
 			return
 		}
 
-		ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
+		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
 		defer cancel()
 
 		resultChan, err := rmq.Conn.Channel()
